streamlit-projekt/requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
plotly>=5.20
pyyaml>=6.0
python-dotenv>=1.0
